+++ src/main/java/model/dao/computer/ComputerDAOImpl.java
@@ -6,9 +6,10 @@ import model.Page;
 import model.company.Company;
 import model.computer.Computer;
 import model.dao.DAOException;
-import model.dao.DAOFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import utils.Utils;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -23,13 +24,11 @@ import java.util.List;
 public class ComputerDAOImpl implements ComputerDAO {
 
     private Logger logger = LoggerFactory.getLogger(ComputerDAOImpl.class);
-    private DAOFactory daoFactory;
+    @Autowired
+    private com.zaxxer.hikari.HikariDataSource dataSource;
 
     public static final String DELETE_SUCCES = "Computers are deleted";
 
-    public void setDaoFactory(DAOFactory daoFactory) {
-        this.daoFactory = daoFactory;
-    }
 
     /**
      * Get the number of Computer.
@@ -38,11 +37,11 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public int count() throws DAOException {
         logger.info("Count computers");
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         ResultSet r = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement("SELECT COUNT(*) FROM computer");
             r = s.executeQuery();
             int result = -1;
@@ -53,7 +52,7 @@ public class ComputerDAOImpl implements ComputerDAO {
         } catch (SQLException e) {
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s, r);
+            Utils.close(conn, s, r);
         }
     }
 
@@ -78,11 +77,11 @@ public class ComputerDAOImpl implements ComputerDAO {
     public Page<Computer> getPage(int page, int sizePage, String order) throws DAOException {
         logger.info("Get page " + page + ", computers of " + sizePage);
         String orderBy = getOrder(order);
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         ResultSet r = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement("SELECT c1.id, c1.name, introduced, discontinued , c2.id, c2.name"
                     + " FROM computer c1"
                     + " LEFT JOIN company c2 ON c1.company_id = c2.id"
@@ -98,7 +97,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s, r);
+            Utils.close(conn, s, r);
         }
     }
 
@@ -130,11 +129,11 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public Page<Computer> getPage(String search) throws DAOException {
         logger.info("Get Search computers : " + search);
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         ResultSet r = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement("SELECT c1.id, c1.name, introduced, discontinued , c2.id, c2.name"
                     + " FROM computer c1"
                     + " LEFT JOIN company c2 ON c1.company_id = c2.id"
@@ -150,7 +149,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s, r);
+            Utils.close(conn, s, r);
         }
     }
 
@@ -162,11 +161,11 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public Computer getDetails(int id) throws DAOException {
         logger.info("Get computer : " + id);
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         ResultSet r = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
 
             s = conn.prepareStatement("SELECT c1.id, c1.name, introduced, discontinued , c2.id, c2.name"
                     + " FROM computer c1"
@@ -184,7 +183,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s, r);
+            Utils.close(conn, s, r);
         }
     }
 
@@ -239,11 +238,11 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public Computer create(Computer computer) throws DAOException {
         logger.info("Create computer : " + computer);
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         ResultSet generatedKeys = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement(
                     "Insert into computer (name,company_id,introduced,discontinued) values (?,?,?,?)"
                     , Statement.RETURN_GENERATED_KEYS);
@@ -280,7 +279,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s, generatedKeys);
+            Utils.close(conn, s, generatedKeys);
         }
     }
 
@@ -292,10 +291,10 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public boolean update(Computer modifiedComputer) throws DAOException {
         logger.info("Update a computer : " + modifiedComputer);
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement(
                     "UPDATE computer SET name = ?, company_id = ?, introduced = ?, discontinued = ? WHERE ID = ?"
                     , Statement.RETURN_GENERATED_KEYS);
@@ -319,7 +318,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s);
+            Utils.close(conn, s);
         }
     }
 
@@ -331,10 +330,10 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public String delete(int id) throws DAOException {
         logger.info("Delete a computer : " + id);
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement("DELETE FROM computer where id = ?");
             s.setInt(1, id);
 
@@ -348,7 +347,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s);
+            Utils.close(conn, s);
         }
     }
 
@@ -360,10 +359,10 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public String delete(List<Integer> listId) throws DAOException {
         logger.info("Delete computers : " + listId);
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             String prepareDelete = prepareDelete(listId.size());
             s = conn.prepareStatement("DELETE FROM computer where id in (" + prepareDelete + ")");
             for (int i = 0; i < listId.size(); i++) {
@@ -380,7 +379,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s);
+            Utils.close(conn, s);
         }
     }
 
@@ -406,11 +405,11 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public void deleteLast() throws DAOException {
         logger.info("Delete last computer");
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         try {
             int id = getLastId();
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement("DELETE FROM computer WHERE id = ?");
             s.setInt(1, id);
 
@@ -422,7 +421,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s);
+            Utils.close(conn, s);
         }
     }
 
@@ -472,11 +471,11 @@ public class ComputerDAOImpl implements ComputerDAO {
         if (last) {
             ordre = " DESC ";
         }
-        Connection conn;
+        Connection conn = null;
         PreparedStatement s = null;
         ResultSet r = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement(
                     "SELECT id FROM computer ORDER BY id" + ordre);
             r = s.executeQuery();
@@ -489,7 +488,7 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s, r);
+            Utils.close(conn, s, r);
         }
     }
 
@@ -500,10 +499,13 @@ public class ComputerDAOImpl implements ComputerDAO {
      */
     public void deleteIdCompany(int id) throws DAOException {
         logger.info("Delete computer of the company : " + id);
-        Connection conn;
+        if (id <= 0) {
+            throw new DAOException("id.invalid");
+        }
+        Connection conn = null;
         PreparedStatement s = null;
         try {
-            conn = daoFactory.get();
+            conn = dataSource.getConnection();
             s = conn.prepareStatement("DELETE FROM computer where company_id = ?");
             s.setInt(1, id);
             s.executeUpdate();
@@ -511,7 +513,32 @@ public class ComputerDAOImpl implements ComputerDAO {
             e.printStackTrace();
             throw new DAOException(e);
         } finally {
-            daoFactory.close(s);
+            Utils.close(conn, s);
+        }
+    }
+
+    /**
+     * Delete all computer of one company.
+     * @param id the id of the company
+     * @param conn the connection with transaction
+     * @throws DAOException id SQL bug
+     */
+    public void deleteIdCompany(Connection conn, int id) throws DAOException {
+        logger.info("Delete computer of the company : " + id);
+        if (conn == null) {
+            throw new DAOException("connection.null");
+        }
+        if (id <= 0) {
+            throw new DAOException("id.invalid");
+        }
+        PreparedStatement s;
+        try {
+            s = conn.prepareStatement("DELETE FROM computer where company_id = ?");
+            s.setInt(1, id);
+            s.executeUpdate();
+        } catch (SQLException e) {
+            e.printStackTrace();
+            throw new DAOException(e);
         }
     }
 }
